Implemented modal--success
After successful sending the question, modal shows in the middle of the page informing that this operation was successful, it blurs the background and doesn't allow to scroll the page untill it is on the top. After clicking OK button the modal disappears and the background goes back to normal.